Deduplicate storage tool tests
