+++ tests/unit/storage-tools.test.js
@@ -1,7 +1,6 @@
 /* global describe, test, expect, beforeEach */
 import { jest } from "@jest/globals";
 
-// Mock all dependencies
 jest.mock("../../src/popup/communication.js", () => ({
   send: jest.fn(),
 }));
@@ -10,42 +9,20 @@ jest.mock("../../src/popup/messages.js", () => ({
   showSuccess: jest.fn(),
 }));
 
-// chrome API will be mocked in beforeEach
-
-// Mock URL and Blob APIs
-globalThis.URL = class {
-  constructor() {
-    this.origin = "https://example.com";
-  }
-};
-globalThis.URL.createObjectURL = jest.fn();
-globalThis.URL.revokeObjectURL = jest.fn();
-
-globalThis.Blob = jest.fn();
-
-// Mock DOM APIs
-globalThis.document = {
-  createElement: jest.fn(),
-  body: {
-    appendChild: jest.fn(),
-    removeChild: jest.fn(),
-  },
-};
-
 import {
   exportLocalStorage,
   importLocalStorageFromText,
 } from "../../src/popup/storage-tools.js";
 import { send } from "../../src/popup/communication.js";
+import { showSuccess } from "../../src/popup/messages.js";
 
 describe("storage tools export/import", () => {
   beforeEach(() => {
+    document.body.innerHTML = "";
     jest.clearAllMocks();
     globalThis.chrome = {
       tabs: {
-        query: jest.fn().mockResolvedValue([
-          { url: "https://example.com" },
-        ]),
+        query: jest.fn().mockResolvedValue([{ url: "https://example.com" }]),
       },
     };
   });
@@ -57,45 +34,54 @@ describe("storage tools export/import", () => {
   test("exportLocalStorage triggers download", async () => {
     send.mockResolvedValue({ foo: "bar" });
 
-    const mockElement = {
-      href: "",
-      download: "",
-      click: jest.fn(),
-    };
+    let createUrlMock;
+    if (URL.createObjectURL) {
+      createUrlMock = jest
+        .spyOn(URL, "createObjectURL")
+        .mockReturnValue("blob:url");
+    } else {
+      URL.createObjectURL = jest.fn(() => "blob:url");
+      createUrlMock = URL.createObjectURL;
+    }
+    if (!URL.revokeObjectURL) {
+      URL.revokeObjectURL = jest.fn();
+    } else {
+      jest.spyOn(URL, "revokeObjectURL");
+    }
 
-    // Mock document.createElement to return our mock element
-    const originalCreateElement = globalThis.document.createElement;
-    globalThis.document.createElement = jest.fn().mockReturnValue(mockElement);
+    const clickMock = jest.fn();
+    const origCreate = document.createElement;
+    document.createElement = jest.fn((tag) => {
+      const el = origCreate.call(document, tag);
+      if (tag === "a") {
+        el.click = clickMock;
+      }
+      return el;
+    });
 
-    // Mock document.body methods to not use real DOM
-    globalThis.document.body.appendChild = jest.fn();
-    globalThis.document.body.removeChild = jest.fn();
-
-    globalThis.URL.createObjectURL.mockReturnValue("blob:url");
-    globalThis.Blob.mockImplementation((content, options) => ({
-      content,
-      options,
-    }));
+    const appendSpy = jest.spyOn(document.body, "appendChild");
+    const removeSpy = jest.spyOn(document.body, "removeChild");
 
     await exportLocalStorage();
 
     expect(send).toHaveBeenCalledWith("getLocalStorage");
-    expect(globalThis.URL.createObjectURL).toHaveBeenCalled();
-    expect(mockElement.click).toHaveBeenCalled();
-    expect(globalThis.URL.revokeObjectURL).toHaveBeenCalledWith("blob:url");
-    expect(globalThis.document.body.appendChild).toHaveBeenCalledWith(
-      mockElement
-    );
-    expect(globalThis.document.body.removeChild).toHaveBeenCalledWith(
-      mockElement
-    );
+    expect(createUrlMock).toHaveBeenCalled();
+    expect(clickMock).toHaveBeenCalled();
+    expect(appendSpy).toHaveBeenCalled();
+    expect(removeSpy).toHaveBeenCalled();
+    expect(URL.revokeObjectURL).toHaveBeenCalledWith("blob:url");
+    expect(showSuccess).toHaveBeenCalled();
 
-    // Restore original
-    globalThis.document.createElement = originalCreateElement;
+    if (createUrlMock.mockRestore) createUrlMock.mockRestore();
+    if (URL.revokeObjectURL.mockRestore) URL.revokeObjectURL.mockRestore();
+    appendSpy.mockRestore();
+    removeSpy.mockRestore();
+    document.createElement = origCreate;
   });
 
   test("importLocalStorageFromText sends data", async () => {
     await importLocalStorageFromText('{"a":1}');
     expect(send).toHaveBeenCalledWith("setLocalStorage", { data: { a: 1 } });
+    expect(showSuccess).toHaveBeenCalled();
   });
 });
